Refactor: spilt to Client and BaseClient
